Split static RuntimeSystem into RuntimeAPI and instance specific RuntimeSystem.
