+++ Src/ILGPU/Resources/RuntimeErrorMessages.resx
@@ -135,6 +135,9 @@
   <data name="InvalidCodeGenerationOperation1" xml:space="preserve">
     <value>Invalid code generation operation:\n{0}</value>
   </data>
+  <data name="InvalidConcurrentModification" xml:space="preserve">
+    <value>Concurrent modification detected</value>
+  </data>
   <data name="InvalidCustomGroupSize" xml:space="preserve">
     <value>Invalid custom group size &gt; 0 in a grouped-index scenario</value>
   </data>
